+++ biopytools/merge_deepbsa/__init__.py
@@ -1,0 +1,13 @@
+"""
+合并Windows版DeepBSA结果|Merge Windows DeepBSA Results
+"""
+
+from .main import WindowsDeepbsaMerger
+from .config import MergeDeepbsaConfig
+
+__version__ = "1.0.0"
+
+__all__ = [
+    'WindowsDeepbsaMerger',
+    'MergeDeepbsaConfig'
+]